backend-django/backend/users/urls.py
"""
User URLs for the Fashion Hub project.
"""

from django.urls import path, include
from rest_framework.routers import DefaultRouter
from rest_framework_simplejwt.views import TokenRefreshView

from .views import UserViewSet, AddressViewSet, RegisterView, CustomTokenObtainPairView

# Create a router and register viewsets
router = DefaultRouter()
router.register(r'', UserViewSet, basename='user')
router.register(r'addresses', AddressViewSet, basename='address')

# URL patterns
urlpatterns = [
    # Include router URLs
    path('', include(router.urls)),
    
    # Authentication endpoints
    path('register/', RegisterView.as_view(), name='register'),
    path('token/', CustomTokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]
