backend-django/requirements.txt
Django
djangorestframework
psycopg2-binary
python-decouple
djangorestframework-simplejwt
django-elasticsearch-dsl
elasticsearch
gunicorn
Pillow>=10.0.0
drf-spectacular
drf-nested-routers
django-filter
requests